Add current theme to fastfetch config (#889)
* Turn off tte until the package is fixed

* Turn off auto-launching screensaver until tte is fixed

* Add current theme to fastfetch config

fastfetch has a module for displaying colors, which is often used by
default. It shows the current colors in use by the terminal. Given that
themes are core to Omarchy, I thought it made sense to show the current
theme in fastfetch. I've also added a compressed version of the palette
(similar to the circle config of the colors module).

* Add migration

---------

Co-authored-by: David Heinemeier Hansson <david@hey.com>